% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{add_emr_job_type}
\alias{add_emr_job_type}
\title{add_emr_job_type}
\usage{
add_emr_job_type(df, position_number_col_name = "Position Number",
  suffix_col_name = "Suffix", mus_col_name = "MUS",
  ecls_col_name = "PEAEMPL ECLS")
}
\arguments{
\item{df}{the dataframe containing the necessary columns and to which the new
column will be appended.}

\item{position_number_col_name}{the string containing the name of the
Position Number column}

\item{suffix_col_name}{the string containing the name of the
Suffix column}

\item{mus_col_name}{the string containing the name of the
MUS Contract Indicator column}
}
\value{
the original dataframe with a newly appended \code{EMRJobType} column
}
\description{
EMR Job Types are aggregated groups of employee classes. These are commonly
used as grouping variables for analyses. Job type derived from Position #,
Suffix, and MUS Contract indicator.
}
