% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{add_emr_orgs}
\alias{add_emr_orgs}
\title{add_emr_orgs}
\usage{
add_emr_orgs(df, dept_number_col_name, opt_lu_fpath)
}
\arguments{
\item{df}{the dataframe to which the data will be added}

\item{dept_number_col_name}{the string containing the name of the dept number
column}

\item{opt_lu_fpath}{supply if an alternative lookup dataset is needed.
Defaults to use the internally stored \code{emr_org_xwalk} dataframe.}
}
\value{
the original dataframe with an additioanl emr_org column
}
\description{
Function to add a column of data to a dataframe corresponding to
}
