% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation_funcs.R
\name{add_flsa_exmpt_status}
\alias{add_flsa_exmpt_status}
\title{Add a job's FLSA exemption status to a dataframe}
\usage{
add_flsa_exmpt_status(df, ecls_col_name)
}
\arguments{
\item{df}{the dataframe containing the ecls column and to which the new
column will be added.}

\item{ecls_col_name}{The string vector containing the name of the eclass
column to which the FLSA status will be mapped.}
}
\value{
The original dataframe with a new column named \code{FLSA OT Exempt}.
}
\description{
Using a choosen Ecls column, map elcs to FLSA overtime exemption status.
}
