% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation_funcs.R
\name{add_longevity_bonus}
\alias{add_longevity_bonus}
\title{add_longevity_bonus}
\usage{
add_longevity_bonus(df, longevity_date_col, hr_rate_col, assgn_rate_col,
  annual_rate_col, opt_as_of_date)
}
\arguments{
\item{df}{the dataframe to which the new columns will be appended}

\item{longevity_date_col}{The name of the column containing the longevity
date data for each person.}

\item{hr_rate_col}{The name of the column containing the unadjusted hourly
rate of the job}

\item{assgn_rate_col}{The name of the column containing the unadjusted
monthly rate of the job}

\item{annual_rate_col}{The name of the column containing the unadjusted
annual salary of the job}

\item{opt_as_of_date}{An optional date on which to caluclate longevity bonus
rates. Defaults to value found in the \code{date} column if unsupplied.}
}
\value{
The original dataframe with four new columns:
\describe{
\item{years_of_service}{The total years of service as-of the last day of the month of the as-of date}
\item{PercentToBase}{The percent bonus as a double i.e. no bonus corresponds with a PercentToBase == 1. Derived from the LongevityBonus Lookup table}
\item{BaseAndLongHourly}{The Hourly pay rate with longevity bonus}
\item{BaseAndLongAssgn}{The monthly pay rate with longevity bonus}
\item{BaseAndLongAnnual}{The annual salary with longevity bonus}
}
}
\description{
Add a longevity raise amount column depending on the years of service and
emr job type. If the EMRJobType column is not in the dataset, the function will
attempt to add it.
}
