% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{allee_dates_from_fnames}
\alias{allee_dates_from_fnames}
\title{allee_dates_from_fnames}
\usage{
allee_dates_from_fnames(file_list)
}
\arguments{
\item{file_list}{a character vector containing the names of all the txt
files in the folder containing csv}
}
\value{
list of POSIXct dates contained in the input filenames
}
\description{
Extracts and formats the dates contained in the csv file names. The csv files
must be in the form "YYYYMMDD All Employees.txt".
}
