% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_fiscal_year}
\alias{compute_fiscal_year}
\title{compute_fiscal_year}
\usage{
compute_fiscal_year(date)
}
\arguments{
\item{date}{the vector of dates from which to compute fiscal year}
}
\value{
the vector of year integers
}
\description{
computes the fiscal year of a vector of dates based on the Montana State
fiscal calendar (July 1 - June 30)
}
\examples{
dte_1 <- as.Date("2018-01-01")
dte_2 <- as.Date("2018-08-01")
compute_fiscal_year(dte_1)
compute_fiscal_year(dte_2)
}
