% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{create_sheets}
\alias{create_sheets}
\title{create_sheets inserts a new sheet and data to the input workbook. Contains
the code for formatting and styles.}
\usage{
create_sheets(df, wb_active, df_name, opt_header_row)
}
\arguments{
\item{df}{the dataframe to be added}

\item{wb_active}{the workbook into which it will be added}

\item{df_name}{the name of the worksheet}

\item{opt_header_row}{an optional header row, default = 2}
}
\value{
NULL as the activeworkbook is being operated on by reference
}
\description{
create_sheets inserts a new sheet and data to the input workbook. Contains
the code for formatting and styles.
}
