% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{ecls_flsa_exmpt_tbl}
\alias{ecls_flsa_exmpt_tbl}
\title{FLSA Overtime exemption status is applied via nbajobs eclass setting. This
setting enables or disables the ability to enter extra overtime hours into
the web time entry timesheet. The dataset maps the ~40 eclasses to their FLSA
OT exemption status. 'FLSA OT Exempt' is one of three values: T, F, or NA. NA
is typically applied to corner cases such as the 'MG' or Additional
Compensation Eclass. This dataset is principally used in the
\code{add_flsa_exmpt_status}. The individual exempt non-exempt designations
should be reviewed on an annual basis with the HR team to ensure that the
Banner implementation has not changed.}
\format{A dataframe with 43 rows and 2 variables:
\describe{
  \item{Ecls Code}{The two digit abbreviated eclass code. Only the eclass
  applied to NBAJOBS record affects FLSA status}
  \item{FLSA OT Exempt}{A boolean value indicating whether the eclass allows
  overtime hours on webtime entry. Acceptable values include T, F, and NA.}}}
\usage{
ecls_flsa_exmpt_tbl
}
\description{
FLSA Overtime exemption status is applied via nbajobs eclass setting. This
setting enables or disables the ability to enter extra overtime hours into
the web time entry timesheet. The dataset maps the ~40 eclasses to their FLSA
OT exemption status. 'FLSA OT Exempt' is one of three values: T, F, or NA. NA
is typically applied to corner cases such as the 'MG' or Additional
Compensation Eclass. This dataset is principally used in the
\code{add_flsa_exmpt_status}. The individual exempt non-exempt designations
should be reviewed on an annual basis with the HR team to ensure that the
Banner implementation has not changed.
}
\keyword{datasets}
