% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fix_native_org_names}
\alias{fix_native_org_names}
\title{fix_native_org_names}
\usage{
fix_native_org_names(df)
}
\arguments{
\item{df}{a dataframe containing all employees data.}
}
\value{
the input dataframe with certain Org Hierarchy values overwritten/
fixed
}
\description{
This function overwrites certain Org Hierarchy values for oddly formatted
departments in the all employees report.
}
