% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msuopa.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{msuopa: A package for common OPA related data tasks}
\description{
msuopa: A package for common OPA related data tasks
}
\section{Foo functions}{

The foo functions ...
}

