% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{format_allEE_dates}
\alias{format_allEE_dates}
\title{format_allEE_dates}
\usage{
format_allEE_dates(df)
}
\arguments{
\item{df}{dataframe containing the all employees data}
}
\value{
the input dataframe with revised date formats
}
\description{
Properly format dates using the ISO YYYY-MM-DD standard. All Employees report
formats them as character type in the DD-MMM-YYYY format.
}
