% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{format_banner_date}
\alias{format_banner_date}
\title{format_banner_date}
\usage{
format_banner_date(df_col)
}
\arguments{
\item{df_col}{a single column vector containing the banner date/timestamps}
}
\value{
the same column vector with the timestamps dropped and remaining
values loaded as date classes.
}
\description{
reformat a banner date column to one interpretable with the base
\code{as.Date()} functiWon. Banner stores dates in the format YYYY-MM-DD
HH:MM:SS. For our purposes, the timestamp is irrelevant and subsequently
dropped.
}
