% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{format_sht_names}
\alias{format_sht_names}
\title{format_sht_names ensures that the name to be used for an execl table or
worksheet is properly formatted}
\usage{
format_sht_names(name_vec)
}
\arguments{
\item{name_vec}{a string to be used as a worksheet name and/or table name}
}
\value{
the properly formatted string. if no formatting is needed then return
the input parameter string.
}
\description{
Requries the length to be less than or equal to 31 (sheet max) and replaces
all spaces with underscores (table requirement)
}
