% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{fread_allee_csv}
\alias{fread_allee_csv}
\title{read_allee_csv}
\usage{
fread_allee_csv(path, name)
}
\arguments{
\item{path}{the full path the csv file}

\item{name}{the full name of the csv file}
}
\value{
an unnamed dataframe containing the all employees data
}
\description{
given a csv file containing the all employees report data, load it into a
dataframe. Uses data.table's \code{fread} function for performance reasons.
Column types are specified using \code{all_ee_col_types} function.
}
\seealso{
allee_dates_from_fnames, all_ee_col_types, get_all_ee
}
