% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{FY_days_hours}
\alias{FY_days_hours}
\title{Number of working hours per month and per year for all fiscal years from
2000-2033 and for the primary employee types: Classified, Professional,
Faculty (9 month), and Faculty (10 month)}
\format{A dataframe with 34 rows and 15 variables: \describe{
  \item{begin_date}{The first day of the fiscal year}
  \item{end_date}{The last day of the fiscal year}
  \item{FY}{the fiscal year}
  \item{days_cls, days_prof, days_fac9, days_fac10}{The number of working
  days for the type of employee in the given fiscal year}
  \item{hours_cls, hours_prof, hours_fac9, hours_fac10}{The working hours per
  pay period for the type of employee}
  \item{hours_per_month_cls, hours_per_month_prof, hours_per_month_fac9,
  hours_per_month_fac10}{The number of hours in each pay period for the
  type of employee}}}
\source{
Max Thompson's email dated 2/22/17
}
\usage{
FY_days_hours
}
\description{
Number of working hours per month and per year for all fiscal years from
2000-2033 and for the primary employee types: Classified, Professional,
Faculty (9 month), and Faculty (10 month)
}
\keyword{datasets}
