% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{get_access_conn}
\alias{get_access_conn}
\title{get_access_conn}
\usage{
get_access_conn(db_file_path)
}
\arguments{
\item{db_file_path}{A full file path to the access database including the
file name.}
}
\value{
a connection object that can be used to exectute operations on the db
}
\description{
get a connection to local or network stored access db. The connection allows
for operations to be made on the access database.
}
