% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_addr_max_dates}
\alias{get_addr_max_dates}
\title{get_addr_max_dates}
\usage{
get_addr_max_dates(pidm_vec, opt_banner_connection)
}
\arguments{
\item{pidm_vec}{a vector containing pidms of the people who's addresses will
be analyzed}

\item{opt_banner_connection}{an optional active banner connection object
typically derived from the \code{get_banner_conn} function. Useful for
minimizing the number of times that a password needs to be entered.}
}
\value{
return a dataframe containing the 'key' comprised of PIDM Address
  Type and the corresponding max date all address records of that type and
  person
}
\description{
select only the most recent address of each type for each pidm. The banner
table stores each address every entered into the system as a unique record.
}
