% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_address_data}
\alias{get_address_data}
\title{get_address_data}
\usage{
get_address_data(pidm_vec, opt_banner_connection)
}
\arguments{
\item{pidm_vec}{a vector of pidms to use a filter in the underlying sql
query, may have issues with excessively large vectors. Requires further
testing.}

\item{opt_banner_connection}{if a banner connection has already been made,
supply it here. Otherwise, this function will prompt for logon credentials
for a one time use connection.}
}
\value{
a dataframe containing one row per person per address type with
  corresponding address columns
}
\description{
get the most up-to-date mailing and campus addresses for a group of pidms
}
