% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_all_ee}
\alias{get_all_ee}
\title{get_all_ee}
\usage{
get_all_ee(folderpath, most_recent_only = TRUE, opt_start_date, opt_end_date,
  supplement = TRUE)
}
\arguments{
\item{folderpath}{the folderpath containing the csv files.}

\item{most_recent_only}{Boolean. Simplest way to specify loading the most recent csv
file only for performance reasons.}

\item{opt_start_date}{An optional start date if only certain files should be
loaded. If not specified and \code{most_recent_only == TRUE}, then will load
all files found in folderpath.}

\item{opt_end_date}{An optional end date if only certain files should be
loaded. If not specified and \code{most_recent_only == TRUE}, then will load
all files found in folderpath.}

\item{supplement}{Boolean value indicating if the dataframes should have
additional derived columns added. Examples of columns include EMR Job Type,
Fiscal Year, and Longevity Bonus.}
}
\value{
a single dataframe containing one or more all employees reports.
report data can be distinguished by the added 'date' column.
}
\description{
Load one or more All Employee reports from csv source. Optionally, save the
compiled RDS file for faster loading in the future.
}
