% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_alt_org_hierarchy}
\alias{get_alt_org_hierarchy}
\title{get_alt_org_hierarchy}
\usage{
get_alt_org_hierarchy()
}
\value{
the dataframe specified by \code{opa_org_xwalk}
}
\description{
loads the opa alternative organization hierarchy specifying the department,
college, and division rollups for each organization code. Update referenced
excel document as necessary.
}
