% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_banner_snapshots}
\alias{get_banner_snapshots}
\title{get_banner_snapshots}
\usage{
get_banner_snapshots(year, sql_file_name, by = "week", debug_date,
  include_leave = FALSE, opt_banner_conn)
}
\arguments{
\item{year}{calendar year overwhich snapshots will be pulled. For one date only, see
\code{debug_date}}

\item{sql_file_name}{optional file name and path if an alternative sql query
is needed. Defaults to OPA snapshot sql query.}

\item{by}{a character vector containing one of "day", "week", "month",
"quarter", or "year". Alternatively, may be a numeric integer. Used to
determine the date sequence from which the sql queries are built.}

\item{debug_date}{an optional date used to specify a single date to run for
a snapshot. Will supersede any values used in \code{year} or \code{by}
parameters}

\item{include_leave}{an optional boolean value indicating whether or not to
include LWOP and LWOP/WB (with Benefits) Jobs in the snapshot}

\item{opt_banner_conn}{an active banner connection object typically derived
from \code{get_banner_conn()}. If not provided will create a temp connection}
}
\value{
a list of snapshot query returns. The names of the list items specify
the date on which query is set.
}
\description{
get_banner_snapshots
}
