% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_budget_org_hierarchy}
\alias{get_budget_org_hierarchy}
\title{Organization hierarchy based on job funding sources.

Determine organization(s) providing the highest percent of a job's funding.
Using this as the primary 'Department', crosswalk appropriate 'College' and
'Division' organizations}
\usage{
get_budget_org_hierarchy(df, lbr_dist_org_col_name, lbr_dist_prcnt_col_name)
}
\arguments{
\item{df}{the dataframe to which the budget org hierarchy will be appended}

\item{lbr_dist_org_col_name}{the name of the column containing the
organization code. defaults to value used in All Employees report.}

\item{lbr_dist_prcnt_col_name}{the name of the column containing the
percent funding contributed by the organization code. defaults to value used
in All Employees report.}
}
\value{
a dataframe with a job-key and funding_dept, funding_college,
funding_division
}
\seealso{
\code{get_alt_org_hierarchy}, \code{opa_org_xwalk}
}
