% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integrity_checks.R
\name{get_mus_professor_discrepencies}
\alias{get_mus_professor_discrepencies}
\title{get_mus_professor_discrepencies}
\usage{
get_mus_professor_discrepencies()
}
\value{
a dateframe containing snapshot info for jobs with 'Professor' in
the position title and no MUS checkbox
}
\description{
Check for Professors without the MUS checkbox. Requires a Banner login. Uses the
OPA Banner snapshot from the date called. May later update to check for Rank
records rather than just 'Professor' in the title.
}
