% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_payroll_data}
\alias{get_payroll_data}
\title{get_payroll_data}
\usage{
get_payroll_data(opt_fpath, opt_fpattern)
}
\arguments{
\item{opt_fpath}{an optional parameter if the payroll files are located in a non-default location}

\item{opt_fpattern}{an optional regex pattern specifying which files to load}
}
\value{
a dataframe containing all payroll data
}
\description{
load payroll datafiles into a single dataframe
}
