% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_race_data}
\alias{get_race_data}
\title{get_race_data}
\usage{
get_race_data(unique_pidms, unique_gids, optional_race_file_path,
  optional_return_all_cols = F)
}
\arguments{
\item{unique_pidms}{a character or numeric vector used to filter the returned
dataset.}

\item{unique_gids}{a character or numeric vector used to filter the returned
dataset. Requires a Banner database connection.}

\item{optional_race_file_path}{A filepath to the access database containing
Bob's Race table}

\item{optional_return_all_cols}{A boolean parameter specifying whether to
return all columns or just the calculated IPEDS value.}
}
\value{
race data for all or a subset population
}
\description{
Get Bob's IPED's Access race table. The path is optional incase the location
changes. The database is currently located at
"X:/IRCommon/RACE/IPEDS_Race.accdb". This table uses PIDM as it's unique key.
If it is necessary to filer or join by GID, the \code{get_pidm_gid_lu}
function is used to join GID to the dataset.
}
