% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{load_sql_qry}
\alias{load_sql_qry}
\title{load_sql_qry}
\usage{
load_sql_qry(file_path, file_name)
}
\arguments{
\item{file_path}{the full folder path containing the file}

\item{file_name}{the full name of the file with extension}
}
\value{
a string containing the sql query
}
\description{
load a properly formatted sql query from a plain text file to be sent to an
oracle db via ROracle::dbSendQuery or an access db via DBI::dbSendQuery
}
