% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{longevity_rates}
\alias{longevity_rates}
\title{The Longevity rate lookup file contains a sequence of integers from -1 to 100
corresponding to the years of service and associated longevity percent bonus.}
\format{A dataframe with 102 rows and 2 variables:
\describe{
  \item{YearsOfService}{An integer between -1 and 100}
  \item{PercentToBase}{A double between 0 and .15}
}}
\usage{
longevity_rates
}
\description{
Each year of service has a Percent column which specifies the amount of bonus given to the employee. The bonus
ranges from 0 (no bonus) to .15 (15% bonus). Has not changed since at least 2005.
}
\keyword{datasets}
