% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{opa_org_xwalk}
\alias{opa_org_xwalk}
\title{The OPA Organization lookup table contains crosswalks from organization codes
to Department, College, and Division names. College and Division are roughly
equivalent to the EMROrg and VPOrg found in the emr_org_xwalk dataset.
Updated as needed (addition of new actively used organization codes). See
\code{get_alt_org_hierarchy} internal function in the msuopa package.}
\format{A dataframe with 485 rows and 4 variables:
\describe{
  \item{orgn_code}{The six digit alpha-numeric organization code defined in
  the FTVORGN validation table}
  \item{Department}{The text description/name associated with the given
  \code{orgn_code} in the FTVORGN banner table}
  \item{College}{The college or unit that is roughly equivalent to the level
  3 organization level}
  \item{Division}{The highest level rollup hierarchy approximately equivalent
  to the level 2 organizations as defined by the business office}}}
\usage{
opa_org_xwalk
}
\description{
The OPA Organization lookup table contains crosswalks from organization codes
to Department, College, and Division names. College and Division are roughly
equivalent to the EMROrg and VPOrg found in the emr_org_xwalk dataset.
Updated as needed (addition of new actively used organization codes). See
\code{get_alt_org_hierarchy} internal function in the msuopa package.
}
\keyword{datasets}
