% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pad_gid}
\alias{pad_gid}
\title{pad_gid}
\usage{
pad_gid(gid_vec)
}
\arguments{
\item{gid_vec}{the vector of gids to add gids}
}
\value{
a vector of gids with appropriately placed preceeding '0's
}
\description{
add dropped zeros to gid values
}
