% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rename_column}
\alias{rename_column}
\title{rename_column}
\usage{
rename_column(df, old_name, new_name)
}
\arguments{
\item{df}{the dataframe containing columns to be renamed}

\item{old_name}{a string containing the name of the column to be renamed}

\item{new_name}{a string containing the new name.}
}
\value{
the same dataset with a renamed column
}
\description{
Quickly rename a column based on it's current name rather than location. This
is helpful in certain instances when a name of a column cannot be determined
in advance.
}
