% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{snapshot_sql_query}
\alias{snapshot_sql_query}
\title{The OPA Employee snapshot sql query in a form to be executed on the Banner
database.}
\format{An object of class \code{character} of length 1.}
\source{
OPA Employee Access database
}
\usage{
snapshot_sql_query
}
\description{
For readable formatting, replace /n with actual new lines characters and
then remove all instances of escape characters ('/'). Requires
'AS_OF_DATE_HERE' to be replaced with a valid character representation of
a date in the form YYYY-MM-DD.
}
\keyword{datasets}
