% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{source_folder_files}
\alias{source_folder_files}
\title{source_folder_files}
\usage{
source_folder_files(folder_path = "./R/")
}
\arguments{
\item{folder_path}{The folder path containing the R files. By default, uses
the ./R/ folder contained in teh working directory}
}
\value{
Success message will be printed to terminal
}
\description{
Source all files in a given folder for the current R session. Any file
ending with *.R will be sourced.
}
