% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{supplement_all_ee}
\alias{supplement_all_ee}
\title{supplement_all_ee}
\usage{
supplement_all_ee(df)
}
\arguments{
\item{df}{the all employees report with unaltered column header names.}
}
\value{
the original input dataframe with the additional columns
}
\description{
A wrapper for several functions that add additional columns to the all
employees report. Adds EMR Job Type, EMR Org, Longevity Bonus, Full Name,
Job Key, Job Date Key, and fiscal year
}
\seealso{
add_emr_job_type, add_emr_orgs, add_longevity_bonus
}
