% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{term_code_lu}
\alias{term_code_lu}
\title{The termination reason codes, description and boolean indicator if the
termination was voluntary by the employee}
\format{A dataframe with 16 rows and three variables: \describe{
  \item{TermReas}{The two digit termination code applied to an individual.
  NOT a job term code reason}
  \item{Desc}{A textual description of the term code}
  \item{IsVoluntary}{Boolean value indicating if the reason is employee
  voluntary}}}
\usage{
term_code_lu
}
\description{
The termination reason codes, description and boolean indicator if the
termination was voluntary by the employee
}
\keyword{datasets}
