% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{write_list_report}
\alias{write_list_report}
\title{write_list_report}
\usage{
write_list_report(df_list, output_name_path)
}
\arguments{
\item{df_list}{the list of dataframes}

\item{output_name_path}{the full name of the folder and file name to be
exported}
}
\description{
Export a list of dataframes to an excel workbook. Each sheet contains a
dataframe in the list. The list should be named. If not, will be given
defaults df_1, df_2, ... df_n.
}
