% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{write_report}
\alias{write_report}
\title{write_report}
\usage{
write_report(df, output_full_path_fname, opt_sheetName, opt_header_title)
}
\arguments{
\item{df}{the dataframe to be output}

\item{output_full_path_fname}{the full name of the path and file name}

\item{opt_sheetName}{the name of the sheet and table will}

\item{opt_header_title}{the name of the header (only visible when printed)}
}
\description{
Export a dataframe to an excel and csv file. Typically used to share
aggregated raw data.
}
