% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_ftvorgn_data}
\alias{get_ftvorgn_data}
\title{get_ftvorgn_data}
\usage{
get_ftvorgn_data(opt_banner_conn)
}
\arguments{
\item{opt_banner_conn}{an active banner connection object typically derived
from \code{get_banner_conn()}. If not provided will create a temp connection}
}
\value{
a dataframe containing all FTVORGN table variables. Depreciated rows
are removed leaving only the most recent record if more than one record
exists for a single ftvorgn organization code
}
\description{
ftvorgn data is used to calculate organization hierarchy via the logical
structure of the organization codes. This validation table is maintained by
the finance team and regularly updated with new organization codes and org.
titles. For this reason, it is recommended to use this function to pull the
most up-to-date ftvorgn table data directly from Banner. Requires Banner
logon credentials.
}
