% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_principle_datasets.R
\name{get_pidm_gid_lu}
\alias{get_pidm_gid_lu}
\title{get_pidm_gid_lu}
\usage{
get_pidm_gid_lu(opt_banner_conn, opt_pidm_vec, opt_gid_vec)
}
\arguments{
\item{opt_banner_conn}{if a banner connection has already been made, supply
it here. Otherwise, this function will prompt for logon credentials for a
one time use connection.}

\item{opt_pidm_vec}{use this optional parameter to filter the underlying sql
query. Useful for time-sensitive applications.}

\item{opt_gid_vec}{use this optional parameter to filter the underlying sql
query. Useful for time-sensitive applications.}
}
\value{
a two column dataframe containing gids and corresponding pidms
}
\description{
get a dataset containing gids and their corresponding pidms. This is pulled
directly from banner. If a banner connection is not possible, see OPA's
employee snapshot files. This dataset is comprehensive for every student and
employee that has ever worked on campus while banner has been implemented
}
