% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{write_report}
\alias{write_report}
\title{write_report}
\usage{
write_report(df, output_name_path, sheetName, header_title)
}
\arguments{
\item{df}{the dataframe to be output}

\item{output_name_path}{the full name of the folder and file name to be
exported}

\item{sheetName}{The string name of the excel sheet}

\item{header_title}{The string to be placed in the header title when printing
the sheet}
}
\description{
Export a dataframe to an excel and csv file. Typically used to share
aggregated raw data.
}
